wrap up with only fp
