+++ tests/e2e/types/datagen.go
@@ -1,0 +1,47 @@
+package types
+
+import (
+	"math/rand"
+	"testing"
+
+	"github.com/babylonchain/babylon/testutil/datagen"
+	bstypes "github.com/babylonchain/babylon/x/btcstaking/types"
+	zctypes "github.com/babylonchain/babylon/x/zoneconcierge/types"
+	"github.com/stretchr/testify/require"
+)
+
+func NewBTCStakingPacketData(packet *bstypes.BTCStakingIBCPacket) *zctypes.ZoneconciergePacketData {
+	return &zctypes.ZoneconciergePacketData{
+		Packet: &zctypes.ZoneconciergePacketData_BtcStaking{
+			BtcStaking: packet,
+		},
+	}
+}
+
+func GenIBCPacket(t *testing.T, r *rand.Rand) *zctypes.ZoneconciergePacketData {
+	// generate a finality provider
+	fpBTCSK, _, err := datagen.GenRandomBTCKeyPair(r)
+	require.NoError(t, err)
+	fpBabylonSK, _, err := datagen.GenRandomSecp256k1KeyPair(r)
+	require.NoError(t, err)
+	fp, err := datagen.GenRandomCustomFinalityProvider(r, fpBTCSK, fpBabylonSK, "consumer-id")
+	require.NoError(t, err)
+
+	packet := &bstypes.BTCStakingIBCPacket{
+		NewFp: []*bstypes.NewFinalityProvider{
+			// TODO: fill empty data
+			&bstypes.NewFinalityProvider{
+				// Description: fp.Description,
+				Commission: fp.Commission.String(),
+				// BabylonPk:  fp.BabylonPk,
+				BtcPkHex: fp.BtcPk.MarshalHex(),
+				// Pop:        fp.Pop,
+				ConsumerId: fp.ConsumerId,
+			},
+		},
+		ActiveDel:   []*bstypes.ActiveBTCDelegation{},
+		SlashedDel:  []*bstypes.SlashedBTCDelegation{},
+		UnbondedDel: []*bstypes.UnbondedBTCDelegation{},
+	}
+	return NewBTCStakingPacketData(packet)
+}
